requirements.txt
python-redmine
pyyaml
